+++ src/flask/templating.py
@@ -37,7 +37,7 @@ def _default_template_ctx_processor() -> dict[str, t.Any]:
 
 
 class Environment(BaseEnvironment):
-    """Works like a regular Jinja2 environment but has some additional
+    """Works like a regular Jinja environment but has some additional
     knowledge of how Flask's blueprint works so that it can prepend the
     name of the blueprint to referenced templates if necessary.
     """