+++ CHANGES.rst
@@ -9,6 +9,17 @@ Unreleased
     can be used without parentheses. :issue:`5729`
 
 
+Version 3.1.2
+-------------
+
+Released 2025-08-19
+
+-   ``stream_with_context`` does not fail inside async views. :issue:`5774`
+-   When using ``follow_redirects`` in the test client, the final state
+    of ``session`` is correct. :issue:`5786`
+-   Relax type hint for passing bytes IO to ``send_file``. :issue:`5776`
+
+
 Version 3.1.1
 -------------
 
@@ -1376,7 +1387,7 @@ Released 2011-09-29, codename Rakija
     of Flask itself and no longer of the test client. This cleaned up
     some internal logic and lowers the odds of runaway request contexts
     in unittests.
--   Fixed the Jinja2 environment's ``list_templates`` method not
+-   Fixed the Jinja environment's ``list_templates`` method not
     returning the correct names when blueprints or modules were
     involved.
 
@@ -1462,7 +1473,7 @@ Released 2010-12-31
 
 -   Fixed an issue where the default ``OPTIONS`` response was not
     exposing all valid methods in the ``Allow`` header.
--   Jinja2 template loading syntax now allows "./" in front of a
+-   Jinja template loading syntax now allows "./" in front of a
     template load path. Previously this caused issues with module
     setups.
 -   Fixed an issue where the subdomain setting for modules was ignored